refactor common components to ui package
