+++ src/api/data/CypherDataTable.js
@@ -5,8 +5,8 @@ import _ from 'lodash';
 import { Grid } from 'semantic-ui-react';
 import ColumnSelector from '../../components/data/ColumnSelector';
 import uuid from 'uuid';
-import NodeLabel from '../../components/NodeLabel';
-import Spinner from '../../components/Spinner';
+import NodeLabel from '../../components/ui/NodeLabel';
+import Spinner from '../../components/ui/Spinner';
 import neo4j from '../driver';
 import sentry from '../sentry/index';
 import CSVDownload from './download/CSVDownload';
