+++ src/components/performance/GCMonitor.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import CypherDataTable from '../data/CypherDataTable';
 import queryLibrary from '../../api/data/queries/query-library';
-import Explainer from '../Explainer';
+import Explainer from '../ui/Explainer';
 
 class GCMonitor extends Component {
     query = queryLibrary.JMX_GARBAGE_COLLECTOR.query;
