+++ src/components/performance/SystemLoad.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import CypherTimeseries from '../timeseries/CypherTimeseries';
 import uuid from 'uuid';
 import queryLibrary from '../../api/data/queries/query-library';
-import Explainer from '../Explainer';
+import Explainer from '../ui/Explainer';
 
 class SystemLoad extends Component {
     state = {
