+++ src/components/ui/Explainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Popup, Icon } from 'semantic-ui-react'
-import kb from '../api/knowledgebase';
+import kb from '../../api/knowledgebase';
 
 export default class Explainer extends Component {
     getContent() {
