+++ src/diagnostic/GeneratePackage.js
@@ -1,18 +1,16 @@
 import React, { Component } from 'react';
 import { Button, Icon, Tab, Checkbox, Message } from 'semantic-ui-react';
+import Spinner from '../components/ui/Spinner';
 import uuid from 'uuid';
+import status from '../api/status/index';
 import moment from 'moment';
-
-import advisor from '../../api/diagnostic/advisor/index';
-import collection from '../../api/diagnostic/collection/index';
-import sentry from '../../api/sentry/index';
-import status from '../../api/status/index';
-
-import Spinner from '../ui/Spinner';
 import Advisor from './advisor/Advisor';
 import ConfigurationDiff from './ConfigurationDiff';
-import hoc from '../higherOrderComponents';
-import JSONDownload from '../data/download/JSONDownload';
+import advisor from './advisor/index';
+import collection from './collection/index';
+import sentry from '../sentry/index';
+import hoc from '../components/higherOrderComponents';
+import JSONDownload from '../components/data/download/JSONDownload';
 
 const UPLOAD_DIAGNOSTICS_BY_DEFAULT = false;
